// Llamapay_Frontendteest/src/app/models/user.ts
export class User {
    userId: number = 0;
    nameUser: string = "";
    lastnameUser: string="";
    emailUser: string="";
    birthdayUser: Date=new Date(Date.now());
    registrationDateUser: Date=new Date(Date.now());
    username: string = "";
    password: string = "";
    enabled: boolean = true;
}